Add exec_depend on python-cbor
rosbridge_library does not depend on this package because it is not
available on Fedora: https://github.com/RobotWebTools/rosbridge_suite/issues/375

The workaround for rosbridge_library was to include the pure Python
implementation of the cbor package in rosbridge_library.  Using the
system package with its native extensions makes encoding binary
arrays roughly twice as fast, so we install it here as a workaround
for the workaround.